examples/libero/requirements_py310_fixed.txt
# 从 examples/libero/requirements.txt 修改而来，适配 Python 3.10
absl-py==2.1.0
certifi==2024.12.14
charset-normalizer==3.4.0
cycler==0.12.1
docstring-parser==0.16
etils==1.3.0
eval-type-backport==0.2.0
evdev==1.7.1
fonttools==4.55.3
glfw==1.12.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
kiwisolver==1.4.7
llvmlite==0.41.1  # 更新：0.36.0 -> 0.41.1 (支持 Python 3.10)
markdown-it-py==3.0.0
matplotlib==3.5.3
mdurl==0.1.2
mujoco==3.2.3
numba==0.58.1  # 更新：0.53.1 -> 0.58.1 (配合新的 llvmlite)
numpy==1.22.4
opencv-python==4.6.0.66
packaging==24.2
pillow==10.4.0
psutil==6.1.0
pygments==2.18.0
pynput==1.7.7
pyopengl==3.1.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-xlib==0.33
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
robosuite==1.4.1
scipy==1.10.1
setuptools==75.3.0
shtab==1.7.1
six==1.17.0
termcolor==2.4.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.67.1
typeguard==4.4.0
typing-extensions==4.12.2
tyro==0.9.2
urllib3==2.2.3
zipp==3.20.2
